feat(前端): 添加岗位管理页面及创建/删除功能
feat(后端): 实现岗位、简历和技能差距的数据库服务

feat(API): 将岗位、简历和差距API迁移到数据库实现

feat(代理): 实现角色初始化管道的LLM节点

feat(数据库): 添加种子脚本和差距进度字段迁移

+++ backend/src/models/gap.py
@@ -3,7 +3,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import String, Integer, ForeignKey, Text, TIMESTAMP, func
+from sqlalchemy import String, Integer, Float, ForeignKey, Text, TIMESTAMP, func
 from sqlalchemy.dialects.postgresql import UUID, JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -31,7 +31,7 @@ class GapItem(IDMixin, TimestampMixin, Base):
     target_state: Mapped[str | None] = mapped_column(Text, nullable=True, default=None)
     evidence_json: Mapped[dict | None] = mapped_column(JSONB, nullable=True, default=None)
     improvement_plan_json: Mapped[dict | None] = mapped_column(JSONB, nullable=True, default=None)
-    progress: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    progress: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
     status: Mapped[str] = mapped_column(
         String(32), nullable=False, default="open",
     )
